Add a jsdoc comment for safeRename
- One of the tests verifies that the result of renaming and/or moving
matches our expectations.

- Also, collapse tests into one file to control when each mock is active.
